+++ extensions/chrome/popup.js
@@ -1,5 +1,5 @@
 "use strict";
 
 document.getElementById("open-etherscan").addEventListener("click", () => {
-  chrome.tabs.create({ url: "https://etherscan.io/txs" });
+  chrome.tabs.create({ url: "https://etherscan.io/contractsVerified" });
 });